# Conflicts:
#	MBTI-backend/src/main/java/tukorea/MBTIbackend/domain/product/controller/ProductController.java
#	MBTI-backend/src/main/java/tukorea/MBTIbackend/domain/product/dto/ProductResponseDto.java
#	MBTI-backend/src/main/java/tukorea/MBTIbackend/domain/product/repository/ProductRepository.java
#	MBTI-backend/src/main/java/tukorea/MBTIbackend/domain/product/service/ProductService.java

+++ MBTI-backend/src/main/java/tukorea/MBTIbackend/domain/product/controller/ProductController.java
@@ -1,7 +1,6 @@
 package tukorea.MBTIbackend.domain.product.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
@@ -15,9 +14,7 @@ public class ProductController {
 
 
     @GetMapping("/api/product/{productId}")
-    public ProductResponseDto findProductById(@PathVariable Long productId) {
-
+    public ProductResponseDto findProductById(@PathVariable Integer productId) {
         return productService.findProductByPID(productId);
     }
-
 }
